Branche feature/psycopg3-migration : 4 commits.

A. Ajout des deps psycopg[binary]==3.3.3 + psycopg-pool==3.3.0
   (cohabitation transitoire avec psycopg2-binary).
B. Migration core indissociable (un commit) : pool ConnectionPool +
   row_factory=dict_row global, suppression des cursor_factory=
   RealDictCursor sur 14 CLIs, 24 sites Json migrés vers
   psycopg.types.json.Jsonb, 10 appels execute_values → executemany,
   adaptations de typage psycopg3 strict (casts %s::int / %s::text[],
   IN %s → = ANY(%s), ARRAY[%s::int], psycopg.errors.UniqueViolation).
C. Drop psycopg2-binary, derniers imports + docstrings actualisés,
   assert sur fetchone() pour mypy strict.
D. ROADMAP §2.11 marqué clôturé avec détails de la migration.

`grep -rn "psycopg2" --include="*.py"` retourne zéro résultat.
Tests : 336 unit + 574 intégration verts, psycopg2 désinstallé.

+++ infrastructure/repositories/config_repository.py
@@ -62,7 +62,7 @@ class PgConfigRepository:
             """
             UPDATE perimeters
             SET structure_ids = array_append(structure_ids, %s)
-            WHERE id = %s AND NOT structure_ids @> ARRAY[%s]
+            WHERE id = %s AND NOT structure_ids @> ARRAY[%s::int]
             RETURNING id
             """,
             (structure_id, perimeter_id, structure_id),
